refactor: 2단계 피드백 반영 - Points 일급 컬렉션 분리
